--- intelmq/bots/parsers/taichung_city_netflow/__init__.py
@@ -1,1 +1,0 @@
-
