+++ intelmq/bots/parsers/urlvir/parser_ips.py
@@ -1,0 +1,39 @@
+from intelmq.lib import utils
+from intelmq.lib.bot import Bot, sys
+from intelmq.lib.message import Event
+from intelmq.lib.harmonization import DateTime, IPAddress
+
+
+class URLVirIPsParserBot(Bot):
+
+    def process(self):
+        report = self.receive_message()
+
+        if not report.contains("raw"):
+            self.acknowledge_message()
+
+        raw_report = utils.base64_decode(report.value("raw"))
+
+        for row in raw_report.split('\n'):
+
+            row = row.strip()
+            if row == "" or row == "#":
+                continue
+
+            event = Event()
+
+            event.add('source.ip', row, sanitize=True)
+
+            time_observation = DateTime().generate_datetime_now()
+            event.add('time.observation', time_observation, sanitize=True)
+            event.add('classification.type', u'malware')
+            event.add('feed.name', report.value("feed.name"))
+            event.add('feed.url', report.value("feed.url"))
+            event.add('raw', row, sanitize=True)
+
+            self.send_message(event)
+        self.acknowledge_message()
+
+if __name__ == "__main__":
+    bot = URLVirIPsParserBot(sys.argv[1])
+    bot.start()
